+++ src/myble.nim
@@ -1,7 +1,11 @@
 import std/asyncdispatch
-import std/logging
+
+when not defined release:
+  import std/logging
+
 from std/options import get, some
-from std/strutils import strip, join, toLowerAscii, contains, split, multiReplace, parseInt
+from std/strutils import strip, join, toLowerAscii, contains, split,
+                          multiReplace, parseInt
 from std/strformat import fmt
 from std/sugar import collect
 from std/tables import Table, `[]`, `[]=`
@@ -10,7 +14,8 @@ import pkg/db_connector/db_sqlite
 
 import pkg/telebot
 
-from pkg/bibleTools import parseBibleVerses, `$`, inOzzuuBible, BibleVerse, UnknownBook, BibleBook, identifyBibleBook
+from pkg/bibleTools import parseBibleVerses, `$`, inOzzuuBible, BibleVerse,
+                            UnknownBook, BibleBook, identifyBibleBook
 
 const apiSecretFile {.strdefine.} = "secret.key"
 let apiSecret = strip readFile apiSecretFile
@@ -18,6 +23,7 @@ let apiSecret = strip readFile apiSecretFile
 var
   db: DbConn
   bookIds: Table[BibleBook, int]
+  ozzuuBibleTransl: string
 
 const
   tgMsgMaxLen = 4096
@@ -26,22 +32,22 @@ const
 func cleanVerse(verse: string): string =
   var words: seq[string]
   for word in verse.multiReplace({
+    "<pb/>": "",
     "<": " <",
     ".>": "/> "
   }).split " ":
     if word.len > 0 and word[0] != '<':
       words.add word
-  result = words.join " "
+  result = strip words.join " "
 
 proc getBookIds =
-  for row in db.rows(sql"SELECT book_number, short_name FROM books_all"):
+  for row in db.rows(sql"SELECT book_number, short_name FROM books"):
     bookIds[row[1].identifyBibleBook.book] = parseInt row[0]
 
 proc getFromDb(verse: BibleVerse): string =
   if verse.book.book != UnknownBook:
     {.gcsafe.}:
       let bookId = bookIds[verse.book.book]
-    echo bookId
     if verse.verses.len == 1:
       let
         verseNum = verse.verses[0]
@@ -53,7 +59,8 @@ proc getFromDb(verse: BibleVerse): string =
         )
       result = fmt"{verseNum} {cleanVerse text}"
     elif verse.verses.len > 0:
-      let versesQuery = collect(for verse in verse.verses: fmt"verse = {verse}").join " OR "
+      let versesQuery = collect(for verse in
+          verse.verses: fmt"verse = {verse}").join " OR "
       for row in db.rows(
         sql fmt"SELECT verse, text FROM verses WHERE book_number = ? AND chapter = ? AND ({versesQuery})",
         bookId,
@@ -82,9 +89,11 @@ proc inlineHandler(b: Telebot, u: InlineQuery): Future[bool] {.async, gcsafe.} =
       shortBook = false
     )
     res.id = $i
-    res.replyMarkup = some newInlineKeyboardMarkup(@[
-      initInlineKeyboardButton(fmt"Open in Ozzuu Bible", verse.inOzzuuBible)
-    ])
+    {.gcsafe.}:
+      res.replyMarkup = some newInlineKeyboardMarkup(@[
+        initInlineKeyboardButton("Open in Ozzuu Bible",
+            verse.inOzzuuBible ozzuuBibleTransl)
+      ])
     res.inputMessageContent = some InputTextMessageContent(
       res.title &
       "\l\l" &
@@ -93,18 +102,16 @@ proc inlineHandler(b: Telebot, u: InlineQuery): Future[bool] {.async, gcsafe.} =
 
     results.add res
 
-  echo u
-
   discard waitFor b.answerInlineQuery(u.id, results)
 
-proc main(mybibleModule: string; dbUser = "", dbPass = "") {.async.} =
-  var L = newConsoleLogger(fmtStr="$levelname, [$time] ")
-  addHandler(L)
-
+proc main(mybibleModule, ozzuuBibleTranslation: string; dbUser = "", dbPass = "") =
+  when not defined release:
+    var L = newConsoleLogger(fmtStr = "$levelname, [$time] ")
+    addHandler(L)
+  ozzuuBibleTransl = ozzuuBibleTranslation
   db = open(mybibleModule, dbUser, dbPass, "")
 
-  {.gcsafe.}:
-    getBookIds()
+  getBookIds()
 
   let bot = newTeleBot apiSecret
   bot.onInlineQuery inlineHandler
@@ -113,6 +120,4 @@ proc main(mybibleModule: string; dbUser = "", dbPass = "") {.async.} =
 when isMainModule:
   import pkg/cligen
 
-  proc cli(mybibleModule: string; dbUser = "", dbPass = "") =
-    waitFor main(mybibleModule, dbUser, dbPass)
-  dispatch cli 
+  dispatch main